# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Barefoot Nomad API",
    "description": "Barefoot Nomad makes company global travel and accommodation easy and convenient for the strong workforce of savvy members of staff, by leveraging the modern web.",
    "contact": {
      "email": "team-banshee@barefoot-nomad.com"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
...
  "paths": {
...
    "/accomodation/feedback": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Feedback"
        ],
        "summary": "Post a feedback for accomodation",
        "operationId": "postAccomodationFeedback",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "schema": {
              "type": "string"
            },
            "description": "user signup token",
            "required": true
          }
        ],
        "requestBody": {
          "description": "user object",
          "content": {
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/accomodationFeedback"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/accomodationFeedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Accomodation feedback posted successfuly"
          },
          "401": {
            "description": "Unathorized, You did not provide a token"
          },
          "500": {
            "description": "Server error, please try again later"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "accomodationFeedback": {
        "type": "object",
        "properties": {
          "reviews": {
            "type": "string"
          },
          "likeDislike": {
            "type": "string"
          },
          "accomodationId": {
            "type": "string",
            "format": "uuid"
          },
          "tripId": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    }
  }
}